refactor(ado): extract PipelineFormatter + decompose pipeline.go (<500 LOC each)
Step A — Create pipeline_format.go (134 LOC):
- Extract presentation formatting from infrastructure adapter
- De-receiver formatPipelineRunsList and FormatRepositoryItems
- Extract formatPipelineRunDetail from AdoGetPipelineRun inline
- Move formatBranchRef, formatKey, isLower/isUpper/toUpper helpers
- ADR-2026-05 compliant: formatting separated from transport

Step B — Decompose pipeline.go into 3 files:
- pipeline.go (246 LOC): types, arg parsing, mapADOVariables,
  filterAndLimitRuns
- pipeline_crud.go (376 LOC): fetchPipelines, AdoListPipelines,
  AdoCreatePipeline, AdoGetBuildChanges, etc.
- pipeline_runs.go (349 LOC): AdoListPipelineRuns, AdoGetPipelineRun,
  AdoRunPipeline, AdoGetBuildTimeline, AdoGetPipelineLogs, etc.

All tests pass with -race, verify_architecture shows zero new edges.
